+++ transfa/__init__.py
@@ -1,5 +1,3 @@
-__version__ = "0.0.1"
-
 # Transfa Python bindings
 # API docs at http://docs.transfapp.com
 # Authors:
